html/dom/documents/dom-tree-accessors/nameditem-fetch-01.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Named items: fetch text/html</title>
<link rel="author" title="bzbarsky" href="mailto:bzbarsky@mit.edu">
<link rel="author" title="ExE Boss" href="https://ExE-Boss.tech">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#dom-htmldocument-nameditem">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
promise_test(function() {
  return /** @type {Promise<Document>} */ (new Promise(function (resolve) {
    var url = 'data:text/html,<html xmlns="http://www.w3.org/1999/xhtml"><form name="x"/></html>';
    var xhr = new XMLHttpRequest();
    xhr.open("GET", url);
    xhr.responseType = "document";
    xhr.onload = function() {
      resolve(xhr.responseXML);
    };
    xhr.send();
  })).then(function (doc) {
    assert_equals(doc.x, doc.getElementsByTagName("form")[0]);
    assert_equals(doc['x'], doc.getElementsByTagName("form")[0]);
  });
}, "If there is one form, it should be returned (name)");
</script>
